+++ CHANGES.md
@@ -1,9 +1,28 @@
 # chill #
 
-### Version 0.0.2 ###
+### 0.1.0
+
+* adds `BijectiveSerializer` and `BijectivePair` with accompanying registrations in `KryoSerializer`
+* adds Enumeration support
+
+### 0.0.5
+
+* ClosureCleaner and Specs (Thanks to the Spark project!)
+
+### 0.0.4
+
+* Added "copy" method to the MeatLocker
+
+### 0.0.3
+
+* Added Manifest serializations
+* Added Traversable serializer
+* Added KryoSerializer object for easy registration
+
+### 0.0.2
 
 * Added TupleSerializers for all tuple types.
 
-### Version 0.0.1 ###
+### 0.0.1
 
 * Added serializers for collection types and MeatLocker.